Move JWT token generation to separated service.
