+++ src/AuthService/Controllers/AuthController.cs
@@ -1,16 +1,13 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using Microsoft.IdentityModel.Tokens;
-using System.IdentityModel.Tokens.Jwt;
-using System.Security.Claims;
-using System.Text;
+using SovietManager.AuthService.Services;
 
 namespace SovietManager.AuthService.Controllers
 {
     [ApiController]
     [Route("api/[controller]")]
-    public class AuthController(IConfiguration configuration) : ControllerBase
+    public class AuthController(IJwtTokenService jwtTokenService) : ControllerBase
     {
-        private readonly IConfiguration _configuration = configuration;
+        private readonly IJwtTokenService _jwtTokenService = jwtTokenService;
 
         [HttpPost("login")]
         public IActionResult Login([FromBody] string userId)
@@ -20,27 +17,8 @@ namespace SovietManager.AuthService.Controllers
                 return Unauthorized();
             }
 
-            var tokenString = GenerateJwtToken(TrimToMaxLength(userId, 36));
+            var tokenString = _jwtTokenService.GenerateJwtToken(TrimToMaxLength(userId, 36));
             return Ok(new { Token = tokenString });
-        }
-
-        private string GenerateJwtToken(string userId)
-        {
-            var tokenHandler = new JwtSecurityTokenHandler();
-            var key = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(_configuration["JWT_SECRET_KEY"]!));
-            var creds = new SigningCredentials(key, SecurityAlgorithms.HmacSha256);
-
-            var tokenDescriptor = new SecurityTokenDescriptor
-            {
-                Subject = new ClaimsIdentity([new Claim("sub", userId.ToString())]),
-                Expires = DateTime.UtcNow.AddHours(1),
-                Issuer = _configuration["JWT_ISSUER"],
-                Audience = _configuration["JWT_AUDIENCE"],
-                SigningCredentials = creds
-            };
-
-            var token = tokenHandler.CreateToken(tokenDescriptor);
-            return tokenHandler.WriteToken(token);
         }
 
         private static string TrimToMaxLength(string input, int maxLength)
