+++ src/AuthService/Services/IJwtTokenService.cs
@@ -1,0 +1,7 @@
+﻿namespace SovietManager.AuthService.Services
+{
+    public interface IJwtTokenService
+    {
+        string GenerateJwtToken(string userId);
+    }
+}
